Add files via upload
Updated to have icons beside host name to show when you are connected to a remote PC.  You can double-click on a "lit up" icon to bring up the uVNC Viewer window that's open for it if it's minimized.  Added a 'Next' button by the search bar, but no code tied to it yet.